refactor event analytics provider
