+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/handlers/streaming/sse/analytics/SseResponseEventDataProvider.java
@@ -23,7 +23,7 @@ import org.apache.synapse.MessageContext;
 /**
  * Data provider for the response events of server sent events.
  */
-public class SseResponseEventDataProvider extends SseEventDataProvider {
+public class SseResponseEventDataProvider extends AsyncAnalyticsDataProvider {
 
     private int responseCode;
